Refactor PopupMenus to UIUtils and add links to StatPanel
