+++ src/main/java/com/flippingutilities/ui/flipping/FlippingItemPanel.java
@@ -38,16 +38,11 @@ import java.awt.Dimension;
 import java.awt.GridLayout;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
-import java.text.Normalizer;
 import java.time.Instant;
-import java.util.Locale;
-import java.util.regex.Pattern;
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
 import javax.swing.JLabel;
-import javax.swing.JMenuItem;
 import javax.swing.JPanel;
-import javax.swing.JPopupMenu;
 import javax.swing.SwingConstants;
 import javax.swing.border.Border;
 import javax.swing.border.CompoundBorder;
@@ -58,17 +53,13 @@ import net.runelite.client.game.ItemManager;
 import net.runelite.client.ui.ColorScheme;
 import net.runelite.client.ui.FontManager;
 import net.runelite.client.util.AsyncBufferedImage;
-import net.runelite.client.util.LinkBrowser;
 import net.runelite.client.util.QuantityFormatter;
-import okhttp3.HttpUrl;
 
 @Slf4j
 public class FlippingItemPanel extends JPanel
 {
 	private static final String NUM_FORMAT = "%,d";
 	private static final String OUTDATED_STRING = "Price is outdated. ";
-	private static final Pattern NONLATIN = Pattern.compile("[^\\w-]");
-	private static final Pattern WHITESPACE = Pattern.compile("[\\s]");
 
 	private static final Border ITEM_INFO_BORDER = new CompoundBorder(
 		BorderFactory.createMatteBorder(0, 0, 0, 0, ColorScheme.DARK_GRAY_COLOR),
@@ -158,20 +149,7 @@ public class FlippingItemPanel extends JPanel
 		itemName.setFont(FontManager.getRunescapeBoldFont());
 		itemName.setPreferredSize(new Dimension(0, 0)); //Make sure the item name fits
 
-		//Opens the item's OSRS Exchange page
-		final JMenuItem openOsrsGe = new JMenuItem("Open in OSRS Exchange");
-		openOsrsGe.addActionListener(e -> LinkBrowser.browse(FlippingItemPanel.buildOsrsExchangeUrl(itemID)));
-
-		//Opens the item's Platinum Tokens page
-		final JMenuItem openPlatinumTokens = new JMenuItem("Open in Platinum Tokens");
-		openPlatinumTokens.addActionListener(e -> LinkBrowser.browse(FlippingItemPanel.buildPlatinumTokensUrl(flippingItem.getItemName())));
-
-		final JPopupMenu popupMenu = new JPopupMenu();
-		popupMenu.setBorder(new EmptyBorder(5, 5, 5, 5));
-		popupMenu.add(openOsrsGe);
-		popupMenu.add(openPlatinumTokens);
-
-		titlePanel.setComponentPopupMenu(popupMenu);
+		titlePanel.setComponentPopupMenu(UIUtilities.createGeTrackerLinksPopup(flippingItem));
 		titlePanel.setBackground(background.darker());
 		titlePanel.add(itemClearPanel, BorderLayout.WEST);
 		titlePanel.add(itemName, BorderLayout.CENTER);
@@ -479,87 +457,6 @@ public class FlippingItemPanel extends JPanel
 				+ ".</html>";
 		}
 		limitLabel.setToolTipText(tooltipText);
-	}
-
-
-	/**
-	 * Builds the url required for opening the OSRS Exchange page for the item.
-	 * <p>
-	 * Example of a lobster's URL:
-	 * http://services.runescape.com/m=itemdb_oldschool/Runescape/viewitem?obj=6416
-	 *
-	 * @param itemId The item to be opened on the Exchange.
-	 * @return Returns the full URL for opening in the browser.
-	 */
-	private static String buildOsrsExchangeUrl(int itemId)
-	{
-		String url = new HttpUrl.Builder()
-			.scheme("http")
-			.host("services.runescape.com")
-			.addPathSegment("m=itemdb_oldschool")
-			.addPathSegment("Runescape")
-			.addPathSegment("viewitem")
-			.addQueryParameter("obj", String.valueOf(itemId))
-			.build()
-			.toString();
-
-		log.info("Opening OSRS Exchange: " + url);
-
-		return url;
-	}
-
-	/**
-	 * This method builds the https://platinumtokens.com (PT) url from the given itemName.
-	 * PT takes a slugged (Dragon dagger(p++) -> dragon-dagger-p-plus-plus) as its item query parameter.
-	 * This method therefore also slugs the item's name, however it's not perfect. There are some continuity errors
-	 * in the slug format used by the site (Rune armour set (sk) WON'T redirect to the item as the (sk) is
-	 * slugged as -sk even though every single other item on the website, with parentheses, are slugged like -s-k).
-	 * Thankfully, this just means the user will be directed to the base URL, so wouldn't be too disruptive for  the user.
-	 * <p>
-	 * Example of an item's url (Dragon dagger(p++)):
-	 * https://platinumtokens.com/item/dragon-dagger-p-plus-plus
-	 *
-	 * @param itemName The item's name to be opened on PT
-	 * @return Returns the URL for the item on PT
-	 */
-	private static String buildPlatinumTokensUrl(String itemName)
-	{
-		//Determine if item name contains parentheses.
-		String[] splitString = itemName.split("\\(");
-		boolean containsParentheses = splitString.length != 1;
-		if (containsParentheses)
-		{
-			//Every character inside parentheses need to be slugged.
-			itemName = splitString[0] + splitString[1].replace("", "-");
-		}
-
-		//'+' is slugged to "plus"
-		itemName = itemName.replace("+", "plus");
-
-		//All whitespaces are replaced with slugs
-		String noWhitespace = WHITESPACE.matcher(itemName).replaceAll("-");
-		//Normalize any characters not expected
-		String normalized = Normalizer.normalize(noWhitespace, Normalizer.Form.NFD);
-		//Remove all remaining parentheses or other symbols and check that we don't have any double slugs.
-		String slug = NONLATIN.matcher(normalized).replaceAll("").replace("--", "-");
-
-		if (containsParentheses)
-		{
-			//If we removed the parentheses earlier, we're guaranteed to have a trailing slug
-			slug = slug.substring(0, slug.length() - 1);
-		}
-
-		//Build the url
-		String url = new HttpUrl.Builder()
-			.scheme("https")
-			.host("platinumtokens.com")
-			.addPathSegment("item")
-			.addPathSegment(slug.toLowerCase(Locale.ENGLISH))
-			.build()
-			.toString();
-
-		log.info("Opening Platinum Tokens: " + url);
-		return url;
 	}
 
 }
